+++ src/refactoring/components/cart/ProductItem.tsx
@@ -1,0 +1,5 @@
+const ProductItem = () => {
+  return;
+};
+
+export default ProductItem;
